@@ -1,14 +1,16 @@
 package com.corgam.cagedmobs.blocks;
 
 import com.corgam.cagedmobs.CagedMobs;
-import com.corgam.cagedmobs.tileEntities.MobCageTE;
-import net.minecraft.client.util.ITooltipFlag;
-import net.minecraft.item.ItemStack;
-import net.minecraft.tileentity.TileEntity;
-import net.minecraft.util.text.ITextComponent;
-import net.minecraft.util.text.TextFormatting;
-import net.minecraft.util.text.TranslationTextComponent;
-import net.minecraft.world.IBlockReader;
+import com.corgam.cagedmobs.blockEntities.MobCageBlockEntity;
+import net.minecraft.ChatFormatting;
+import net.minecraft.core.BlockPos;
+import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.TranslatableComponent;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.TooltipFlag;
+import net.minecraft.world.level.BlockGetter;
+import net.minecraft.world.level.block.entity.BlockEntity;
+import net.minecraft.world.level.block.state.BlockState;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
 
@@ -21,20 +23,20 @@ public class HoppingMobCageBlock extends MobCageBlock{
     }
 
     @Override
-    public TileEntity newBlockEntity(IBlockReader worldIn) {
-        return new MobCageTE(true);
+    public BlockEntity newBlockEntity(BlockPos pos, BlockState state) {
+        return new MobCageBlockEntity(pos, state, true);
     }
 
     @Override
     @OnlyIn(Dist.CLIENT)
-    public void appendHoverText (ItemStack stack, @Nullable IBlockReader worldIn, List<ITextComponent> tooltip, ITooltipFlag flagIn) {
+    public void appendHoverText(ItemStack item, @Nullable BlockGetter getter, List<Component> tooltip, TooltipFlag flag) {
         if(CagedMobs.SERVER_CONFIG.ifHoppingCagesDisabled()){
-            tooltip.add(new TranslationTextComponent("block.cagedmobs.mobcage.hoppingCagesDisabled1").withStyle(TextFormatting.RED));
-            tooltip.add(new TranslationTextComponent("block.cagedmobs.mobcage.hoppingCagesDisabled2").withStyle(TextFormatting.RED));
+            tooltip.add(new TranslatableComponent("block.cagedmobs.mobcage.hoppingCagesDisabled1").withStyle(ChatFormatting.RED));
+            tooltip.add(new TranslatableComponent("block.cagedmobs.mobcage.hoppingCagesDisabled2").withStyle(ChatFormatting.RED));
         }
-        tooltip.add(new TranslationTextComponent("block.cagedmobs.mobcage.mainInfo").withStyle(TextFormatting.GRAY));
-        tooltip.add(new TranslationTextComponent("block.cagedmobs.mobcage.hoppingInfo").withStyle(TextFormatting.GRAY));
-        tooltip.add(new TranslationTextComponent("block.cagedmobs.mobcage.envInfo").withStyle(TextFormatting.GRAY));
-        tooltip.add(new TranslationTextComponent("block.cagedmobs.mobcage.upgrading").withStyle(TextFormatting.GRAY));
+        tooltip.add(new TranslatableComponent("block.cagedmobs.mobcage.mainInfo").withStyle(ChatFormatting.GRAY));
+        tooltip.add(new TranslatableComponent("block.cagedmobs.mobcage.hoppingInfo").withStyle(ChatFormatting.GRAY));
+        tooltip.add(new TranslatableComponent("block.cagedmobs.mobcage.envInfo").withStyle(ChatFormatting.GRAY));
+        tooltip.add(new TranslatableComponent("block.cagedmobs.mobcage.upgrading").withStyle(ChatFormatting.GRAY));
     }
 }
