@@ -1,11 +1,11 @@
 package com.corgam.cagedmobs.items;
 
-import net.minecraft.client.util.ITooltipFlag;
-import net.minecraft.item.ItemStack;
-import net.minecraft.util.text.ITextComponent;
-import net.minecraft.util.text.TextFormatting;
-import net.minecraft.util.text.TranslationTextComponent;
-import net.minecraft.world.World;
+import net.minecraft.ChatFormatting;
+import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.TranslatableComponent;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.TooltipFlag;
+import net.minecraft.world.level.Level;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
 
@@ -19,9 +19,9 @@ public class CookingUpgradeItem extends UpgradeItem{
 
     @Override
     @OnlyIn(Dist.CLIENT)
-    public void appendHoverText (ItemStack stack, @Nullable World worldIn, List<ITextComponent> tooltip, ITooltipFlag flagIn) {
+    public void appendHoverText(ItemStack stack, @Nullable Level worldIn, List<Component> tooltip, TooltipFlag flagIn) {
         super.appendHoverText(stack, worldIn, tooltip, flagIn);
-        tooltip.add(new TranslationTextComponent("item.cagedmobs.cookingupgrade.info").withStyle(TextFormatting.GRAY));
-        tooltip.add(new TranslationTextComponent("item.cagedmobs.upgrades.attach").withStyle(TextFormatting.GRAY));
+        tooltip.add(new TranslatableComponent("item.cagedmobs.cookingupgrade.info").withStyle(ChatFormatting.GRAY));
+        tooltip.add(new TranslatableComponent("item.cagedmobs.upgrades.attach").withStyle(ChatFormatting.GRAY));
     }
 }
