@@ -1,22 +1,21 @@
 package com.corgam.cagedmobs.serializers.mob;
 
 import com.corgam.cagedmobs.CagedMobs;
-import com.corgam.cagedmobs.configs.ServerConfig;
 import com.corgam.cagedmobs.serializers.SerializationHelper;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
-import net.minecraft.entity.EntityType;
-import net.minecraft.item.ItemStack;
-import net.minecraft.item.crafting.IRecipeSerializer;
-import net.minecraft.nbt.CompoundNBT;
-import net.minecraft.network.PacketBuffer;
-import net.minecraft.util.JSONUtils;
-import net.minecraft.util.ResourceLocation;
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.network.FriendlyByteBuf;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.util.GsonHelper;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.crafting.RecipeSerializer;
 import net.minecraftforge.registries.ForgeRegistryEntry;
 
 import java.util.*;
 
-public class MobDataSerializer extends ForgeRegistryEntry<IRecipeSerializer<?>> implements IRecipeSerializer<MobData>{
+public class MobDataSerializer extends ForgeRegistryEntry<RecipeSerializer<?>> implements RecipeSerializer<MobData> {
 
     public static final MobDataSerializer INSTANCE = new MobDataSerializer();
 
@@ -32,16 +31,16 @@ public class MobDataSerializer extends ForgeRegistryEntry<IRecipeSerializer<?>> 
         // Envs
         final Set<String> validEnvs = deserializeEnvsData(id, json);
         // Total grow ticks
-        final int growTicks = JSONUtils.getAsInt(json, "growTicks");
+        final int growTicks = GsonHelper.getAsInt(json, "growTicks");
         // If requires water
         boolean requiresWater = false;
         if(json.has("requiresWater")) {
-            requiresWater = JSONUtils.getAsBoolean(json, "requiresWater");
+            requiresWater = GsonHelper.getAsBoolean(json, "requiresWater");
         }
         // Loot Data
         final List<LootData> results = deserializeLootData(id, json, entityType);
         // Sampler tier
-        final int samplerTier = JSONUtils.getAsInt(json, "samplerTier");
+        final int samplerTier = GsonHelper.getAsInt(json, "samplerTier");
 
         //Error checks
         if (growTicks <= 0){
@@ -55,7 +54,7 @@ public class MobDataSerializer extends ForgeRegistryEntry<IRecipeSerializer<?>> 
     }
 
     @Override
-    public MobData fromNetwork(ResourceLocation id, PacketBuffer buffer) {
+    public MobData fromNetwork(ResourceLocation id, FriendlyByteBuf buffer) {
         try {
             // Entity
             final EntityType<?> entityType = SerializationHelper.deserializeEntityType(id, buffer);
@@ -84,7 +83,7 @@ public class MobDataSerializer extends ForgeRegistryEntry<IRecipeSerializer<?>> 
     }
 
     @Override
-    public void toNetwork(PacketBuffer buffer, MobData recipe) {
+    public void toNetwork(FriendlyByteBuf buffer, MobData recipe) {
         try {
             // Entity
             SerializationHelper.serializeEntityType(buffer, recipe.getEntityType());
@@ -135,7 +134,7 @@ public class MobDataSerializer extends ForgeRegistryEntry<IRecipeSerializer<?>> 
     }
 
     private static ItemStack writeNBTtoItem(String nbtName, String nbtData, ItemStack stack){
-        CompoundNBT nbt = new CompoundNBT();
+        CompoundTag nbt = new CompoundTag();
         nbt.putString(nbtName,nbtData);
         stack.setTag(nbt);
         return stack;
