@@ -1,8 +1,8 @@
 package com.corgam.cagedmobs.serializers.mob;
 
-import net.minecraft.item.crafting.IRecipeType;
+import net.minecraft.world.item.crafting.RecipeType;
 
-public class RecipeTypeMobData implements IRecipeType<MobData> {
+public class RecipeTypeMobData implements RecipeType<MobData> {
     @Override
     public String toString () {
         return "cagedmobs:mob_data";
